% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assessment}
\alias{assessment}
\title{Assessment}
\usage{
assessment(data = NULL, model_dataframe = NULL)
}
\arguments{
\item{data}{Dataframe of variables in hera inter-change format}

\item{model_dataframe}{Dataframe of model_dataframe see `model_dataframe`}
}
\value{
Dataframe of assessments
}
\description{
Run assessment
}
\details{
\code{assessment()} assessment
}
\examples{
assessments <- assessment(demo_data)
}
