% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{bundle an object}
\usage{
bundle(obj, mimetypes = getOption("jupyter.display_mimetypes"), ...)
}
\arguments{
\item{obj}{The object to create representations for}

\item{mimetypes}{Mimetypes to create reprs for. The defaults are defined by the option \code{jupyter.display_mimetypes}. (see: \link[IRdisplay]{IRdisplay-options})}
}
\description{
bundle an object
}
\seealso{
IRdisplay::prepare_mimebundle which this currently wraps around
}
