% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_options.R
\name{cell_options}
\alias{cell_options}
\title{Options for current cell}
\usage{
cell_options(...)
}
\arguments{
\item{...}{For \code{local_options()} and \code{push_options()}, named
values defining new option values. For \code{peek_options()}, strings
or character vectors of option names.}
}
\description{
Options for current cell
}
