% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Classification}
\usage{
classification(data = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of classifications
}
\description{
Model classification.
}
\details{
\code{classification()} classification
}
\examples{
classifications <- classification(demo_data)
}
