% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine}
\usage{
combine(outcome, data)
}
\arguments{
\item{outcome}{The output from prediction or other hera function}

\item{data}{The input data to prediction or other hera function}
}
\description{
Output with all columns from data relevant to assessment name. Either by
sample_id or location_id (depending how output is calculated)
}
