% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completion.R
\name{complete}
\alias{complete}
\title{Code completion}
\usage{
complete(code, cursor_pos = nchar(code))
}
\arguments{
\item{code}{R code to complete}

\item{cursor_pos}{position of the cursor}
}
\value{
a list that contains potential completions as the first item
}
\description{
Code completion
}
