% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert}
\usage{
convert(data, convert_to = "hera", convert_from = "sepa_lims")
}
\arguments{
\item{data}{Data as raw input from a number of sources}

\item{convert_to}{Convert data to `hera` format by default. Currently, a
reverse convert back to the original input format is not possible.}

\item{convert_from}{Specify the structure of the input data. This can be
'sepa' or 'sepa_lims'. 'sepa' is the internal, historic reportable analysis
results structure, 'lims' is the new results structure direct from the lab
info system.}
}
\value{
Dataframe in `hera` structure. See `demo_data`
}
\description{
Convert input data into internal `hera` structure. This allows data from
different sources to be used as input.
}
