% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rict-convertor.R
\name{convert_rict}
\alias{convert_rict}
\title{RICT convertor}
\usage{
convert_rict(data)
}
\arguments{
\item{data}{}
}
\value{
data.frame in RICT input format.
}
\description{
Convert hera dataframe into RICT input dataframe
}
\examples{
\dontrun{
hera:::convert_rict(demo_data)
}
}
