% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-catalogue.R
\name{create_catalogue}
\alias{create_catalogue}
\title{Knit Together Model Functions}
\usage{
create_catalogue()
}
\description{
A low-level function for package developers to update the dataframe of model
functions required for each WFD parameter.
}
\examples{
\dontrun{
catalogue <- hera:::create_catalogue()
usethis::use_data(catalogue, overwrite = TRUE)
}
}
