% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-model-dataframe.R
\name{create_model_dataframe}
\alias{create_model_dataframe}
\title{Knit Together Model Functions}
\usage{
create_model_dataframe()
}
\description{
A low-level function for package developers to update the dataframe of model
functions required for each WFD parameter
}
\examples{
\dontrun{
model_dataframe <- hera:::create_model_dataframe()
usethis::use_data(model_dataframe, overwrite = TRUE)
}
}
