% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Example dataset of WFD data interchange format}
\format{
A data frame with 8477 rows and 40 variables:
\describe{
  \item{analysis_name}{Name of analysis}
  \item{date_taken}{}
  \item{easting}{}
  \item{grid_reference}{}
  \item{latitude}{}
  \item{location_description}{}
  \item{location_id}{}
  \item{longitude}{}
  \item{mean_alkalinity}{}
  \item{northing}{}
  \item{quality_element}{}
  \item{question}{}
  \item{response}{}
  \item{result_id}{}
  \item{sample_id}{}
  \item{season}{}
  \item{standard}{}
  \item{taxon}{}
  \item{taxon_id}{}
  \item{units}{}
  \item{water_body_id}{}
}
}
\source{
Agency sampling data
}
\usage{
demo_data
}
\description{
A dataset containing ecology and chemistry data
}
\keyword{datasets}
