% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-input.R
\docType{data}
\name{demo_input}
\alias{demo_input}
\title{Example dataset of WFD data interchange format}
\format{
A data frame with 8477 rows and 20 variables:
\describe{
  \item{location_id}{location_id}
  \item{location_description}{location_description}
  \item{easting}{easting}
  \item{northing}{northing}
  \item{latitude}{latitude}
  \item{longitude}{longitude}
  \item{date_taken}{date_taken}
  \item{sample_id}{sample_id}
  \item{analysis_name}{analysis_name}
  \item{analysis_repname}{analysis_repname}
  \item{question}{question}
  \item{response}{response}
  \item{units}{units}
  \item{taxon}{taxon}
  \item{taxon_id}{taxon_id}
  \item{mean_alkalinity}{mean_alkalinity}
  \item{result_id}{result_id}
  \item{grid_reference}{grid_reference}
  \item{standard}{standard}
  \item{quality_element}{quality_element}
}
}
\source{
Agency sampling data
}
\usage{
demo_input
}
\description{
A dataset containing ecology and chemistry data
}
\keyword{datasets}
