% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{display_data}
\alias{display_data}
\title{Display data}
\usage{
display_data(data = NULL, metadata = NULL)
}
\arguments{
\item{data}{data to display}

\item{metadata}{potential metadata}
}
\description{
Display data
}
\examples{
\dontrun{
  display_data(mtcars)
}

}
