% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_data.R
\name{get_bulk_data}
\alias{get_bulk_data}
\title{Get Environment Agency demo data}
\usage{
get_bulk_data()
}
\value{
dataframe of demo data for testing
}
\description{
Get Environment Agency demo data
}
\examples{
\dontrun{
data <- get_demo_data()
}
}
