% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(location_id = NULL, take = 10000, date_from = NULL, date_to = NULL)
}
\arguments{
\item{location_id}{Unique ID of location.}

\item{take}{Number of observation to download.}

\item{date_from}{Start of date taken window in string format: "2013-12-31".}

\item{date_to}{End of date taken window in string format: "2015-12-31".}
}
\value{
Data frame
}
\description{
Import data from web services (currently EA only) and convert into standard
format for `hera` regulatory assessment tool.
}
\examples{
\dontrun{
data <- get_data(location_id = 100)
class <- assessment(data)
}
}
