% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera.R
\name{hera}
\alias{hera}
\title{Hera}
\usage{
hera(data = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of classifications
}
\description{
Validate, Predict and Classify.
}
\details{
\code{classification()} classification
}
\examples{
classifications <- classification(demo_data)
}
