% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera-app.R
\name{hera_app}
\alias{hera_app}
\title{Open RICT web app}
\usage{
hera_app()
}
\description{
Open RICT as an interactive shiny web app.
}
\details{
\code{rict_app()} opens RICT as an interactive shiny app.

Using the app:
\enumerate{
 \item Select your required model/area/year type options.
 \item Click the Browse for .csv file button to select and upload a data file.
 \item View and download results.
 }
}
\examples{
\dontrun{
rict_app()
}

}
