% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera-indices.R
\name{hera_indices}
\alias{hera_indices}
\title{Calculate Indices}
\usage{
hera_indices(data)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of predictions
}
\description{
Calculates indices from raw data prior to predictions and classification.
}
\details{
\code{hera_predict()} predicts model indices.
}
\examples{
\dontrun{
indices <- hera_indices(demo_data)
}
}
