% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera_predict.R
\name{hera_predict}
\alias{hera_predict}
\title{Predict hera models}
\usage{
hera_predict()
}
\description{
Predicts hera model indices.
}
\details{
\code{hera_predict()} predicts model indices.
}
\examples{
\dontrun{
hera_predict()
}

}
