% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_datasuite}
\alias{is_datasuite}
\title{Is this a running datasuite jupyter kernel}
\usage{
is_datasuite()
}
\value{
TRUE if the current session is running in a datasuite kernel
}
\description{
Is this a running datasuite jupyter kernel
}
\examples{
is_datasuite()

}
