% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_elara}
\alias{is_elara}
\title{Is this a running Elara jupyter kernel}
\usage{
is_elara()
}
\value{
TRUE if the current session is running in an Elara kernel
}
\description{
Is this a running Elara jupyter kernel
}
\examples{
is_elara()

}
