% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_xeusr}
\alias{is_xeusr}
\title{Is this a running xeusr jupyter kernel}
\usage{
is_xeusr()
}
\value{
TRUE if the current session is running in a xeusr kernel
}
\description{
Is this a running xeusr jupyter kernel
}
