% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-app.R
\name{launch_app}
\alias{launch_app}
\title{Launch hera app}
\usage{
launch_app(new_catalogue = NULL, data = NULL)
}
\arguments{
\item{new_catalogue}{Dataframe of catalogue see `catalogue`}

\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Shiny app
}
\description{
Launches hera shiny app with optional data and `catalogue`.
}
\examples{
\dontrun{
launch_app()
}
}
