% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims-data.R
\docType{data}
\name{lims_data}
\alias{lims_data}
\title{Example dataset of raw LIMS data}
\format{
A data frame with 42 rows and 16 variables:
\describe{
  \item{ORIGINAL_SAMPLE}{ORIGINAL_SAMPLE ID}
  \item{SAMPLE_NUMBER}{SAMPLE_NUMBER}
  \item{SAMPLING_POINT}{SAMPLING_POINT}
  \item{DESCRIPTION}{DESCRIPTION}
  \item{STATUS}{STATUS}
  \item{LOCATION}{LOCATION}
  \item{ANALYSIS}{ANALYSIS}
  \item{REPORTED_NAME}{REPORTED_NAME}
  \item{FORMATTED_ENTRY}{FORMATTED_ENTRY}
  \item{UNITS}{UNITS}
  \item{SAMPLED_DATE}{SAMPLED_DATE}
  \item{TEMPLATE}{TEMPLATE}
  \item{SAMPLED_BY}{SAMPLED_BY}
}
}
\source{
Agency sampling data
}
\usage{
lims_data
}
\description{
Example dataset of raw LIMS data
}
\keyword{datasets}
