% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime_bundle.R
\name{mime_bundle}
\alias{mime_bundle}
\title{bundle an object}
\usage{
mime_bundle(x, mimetypes = mime_types(x), ...)
}
\arguments{
\item{x}{an object}

\item{mimetypes}{mime types}

\item{...}{extra currently unused parameters}
}
\description{
bundle an object
}
\examples{
mime_bundle(letters)

}
\seealso{
IRdisplay::prepare_mimebundle, which does it for objects base R can't
}
