% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mime_bundle.R
\name{mime_types}
\alias{mime_types}
\title{MIME types supported by an object}
\usage{
mime_types(x)
}
\arguments{
\item{x}{an object}
}
\value{
a character vector of its supported mime types
}
\description{
MIME types supported by an object
}
\examples{
mime_types(letters)
mime_types(mtcars)

}
