% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Validation}
\usage{
validation(data = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of validation messages
}
\description{
Validation of input data.
}
\details{
\code{validation()} Validation
}
\examples{
validations <- validation(demo_data)
}
