% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Classification}
\usage{
classification(data = NULL, model_dataframe = NULL)
}
\arguments{
\item{data}{Dataframe of variables in hera inter-change format}

\item{model_dataframe}{Dataframe of model_dataframe see `model_dataframe`}
}
\value{
Dataframe of classifications
}
\description{
Model classification.
}
\details{
\code{classification()} classification
}
\examples{
classifications <- classification(demo_data)
}
