% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine output with all columns from data relevant to assessment name
Either by sample_id or location_id (depending how output is calculated)}
\usage{
combine(output, data, name = NULL)
}
\description{
Combine output with all columns from data relevant to assessment name
Either by sample_id or location_id (depending how output is calculated)
}
