% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\name{Comm}
\alias{Comm}
\title{Comm class}
\description{
Comm class

Comm class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{id of the comm. read only.}

\item{\code{target_name}}{name of the target for this comm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Comm-new}{\code{Comm$new()}}
\item \href{#method-Comm-open}{\code{Comm$open()}}
\item \href{#method-Comm-close}{\code{Comm$close()}}
\item \href{#method-Comm-send}{\code{Comm$send()}}
\item \href{#method-Comm-on_close}{\code{Comm$on_close()}}
\item \href{#method-Comm-on_message}{\code{Comm$on_message()}}
\item \href{#method-Comm-print}{\code{Comm$print()}}
\item \href{#method-Comm-clone}{\code{Comm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-new"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$new(xp, description = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xp}}{external pointer to an instance of the C++ class 'xeus::xcomm'}

\item{\code{description}}{description of this comm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-open"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$open(data = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{metadata}}{metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-close"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$close(data = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{metadata}}{metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-send"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$send(data = NULL, metadata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{metadata}}{metadata}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-on_close"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-on_close}{}}}
\subsection{Method \code{on_close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$on_close(handler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{handler}}{function to call when the comm is closed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-on_message"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-on_message}{}}}
\subsection{Method \code{on_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$on_message(handler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{handler}}{function to call when receiving a message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-print"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
information about the comm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Comm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
