% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.R
\docType{data}
\name{CommManager}
\alias{CommManager}
\title{Comm manager}
\format{
An object of class \code{CommManagerClass} (inherits from \code{R6}) of length 10.
}
\usage{
CommManager
}
\description{
Comm manager
}
\keyword{datasets}
