% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Example dataset of WFD data interchange format}
\format{
A data frame with 8477 rows and 40 variables:
\describe{
  \item{analysis_name}{Name of analysis}
  \item{analysis_repname}{}
  \item{catchment}{}
  \item{catchment_id}{}
  \item{comments}{}
  \item{date_taken}{}
  \item{determinand}{}
  \item{determinand_code}{}
  \item{easting}{}
  \item{hydro_code}{}
  \item{hydro_distance}{}
  \item{latitude}{}
  \item{local_authority}{}
  \item{location_code}{}
  \item{location_description}{}
  \item{longitude}{}
  \item{maitland_code}{}
  \item{mean_alkalinity}{}
  \item{media}{}
  \item{national_grid_reference}{}
  \item{nbn_code}{}
  \item{northing}{}
  \item{purpose}{}
  \item{result}{}
  \item{result_id}{}
  \item{result_number}{}
  \item{result_order}{}
  \item{river_name}{}
  \item{river_number}{}
  \item{sample_id}{}
  \item{sample_number}{}
  \item{sampler_id}{}
  \item{season}{}
  \item{taxon}{}
  \item{test_number}{}
  \item{units}{}
  \item{value}{}
  \item{value_warning}{}
  \item{wfd_purpose}{}
  \item{whitton_code}{}
}
}
\source{
Agency sampling data
}
\usage{
demo_data
}
\description{
A dataset containing ecology and chemistry data
}
\keyword{datasets}
