% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(
  location_id = NULL,
  take = 10000,
  date_from = NULL,
  date_to = NULL,
  dataset = "analytical_results",
  year = NULL,
  water_body_id = "",
  source = "sepa"
)
}
\arguments{
\item{location_id}{Unique ID of location.}

\item{take}{Number of observation to download. For "EA" API services.}

\item{date_from}{Start of date taken window in string format: "2013-12-31".}

\item{date_to}{End of date taken window in string format: "2015-12-31".}

\item{dataset}{Default will get Ecology monitoring data, set to "replocs" to
represent location data for SEPA}

\item{year}{Classification year}

\item{water_body_id}{Water body ID used for replocs table queries.}

\item{source}{Which data source, either "EA or "SEPA". SEPA is internal access
only.}
}
\value{
Data frame
}
\description{
Import data from web services and convert into standard format for `hera`
regulatory assessment tool.
}
\examples{
\dontrun{
data <- get_data(location_id = 1000, source = "ea")
class <- assessment(data)
}
}
