% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera.R
\name{hera}
\alias{hera}
\title{Hera}
\usage{
hera(data = NULL)
}
\arguments{
\item{data}{Dataframe of variables in `hera` inter-change format}
}
\value{
Dataframe of assessments
}
\description{
Validate, Predict and Assess
}
\details{
\code{assessment()} assessment
}
\examples{
assessments <- hera(hera::demo_data)
}
