% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera-app.R
\name{hera_app}
\alias{hera_app}
\title{Open hera web app}
\usage{
hera_app()
}
\description{
Open hera as an interactive shiny web app.
}
\details{
\code{hera_app()} opens hera as an interactive shiny app.
}
\examples{
\dontrun{
hera_app()
}

}
