% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hera-predict.R
\name{hera_predict}
\alias{hera_predict}
\title{Predict hera models}
\usage{
hera_predict(data)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of predictions
}
\description{
Predicts hera model reference indices.
}
\details{
\code{hera_predict()} predicts model indices.
}
\examples{
hera_predictions <- hera_predict(demo_data)
}
