% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Calculate Observed Indices}
\usage{
indices(data, index = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}

\item{index}{Optionally specify which indices to run e.g. c("tdi4", tdi3")}
}
\value{
Dataframe of predictions
}
\description{
Calculates indices from raw data prior to predictions and classification.
}
\details{
\code{indices()} calculates observed indices.
}
\examples{
indices <- indices(demo_data)
}
