% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Calculate Observed Indices}
\usage{
indices(data, name = NULL, catalogue = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}

\item{name}{Name of the assessment to be used}

\item{catalogue}{Dataframe of catalogue see `catalogue`}
}
\value{
Dataframe of predictions
}
\description{
Calculates indices from raw data prior to predictions and classification.
}
\details{
\code{indices()} calculates observed indices.
}
\examples{
indices <- indices(demo_data)
}
