% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict hera models}
\usage{
prediction(data = NULL, name = NULL, model_dataframe = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}

\item{name}{Name of assessment to predict. Default is all possible
assessments.}

\item{model_dataframe}{Dataframe of model_dataframe see `model_dataframe`}
}
\value{
Dataframe of predictions
}
\description{
Predicts hera model reference indices.
}
\details{
\code{prediction()} predicts model indices.
}
\examples{
predictions <- prediction(demo_data)
}
