% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict hera models}
\usage{
prediction(data = NULL, model_dataframe = NULL, combine = TRUE)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}

\item{model_dataframe}{Dataframe of model_dataframe see `model_dataframe`}

\item{combine}{Specify if predictions should be combined within location/sample attributes}
}
\value{
Dataframe of predictions
}
\description{
Predicts hera model reference indices.
}
\details{
\code{prediction()} predicts model indices.
}
\examples{
predictions <- prediction(demo_data)
}
