% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Predict hera models}
\usage{
prediction(data = NULL)
}
\arguments{
\item{data}{Dataframe of variables in WFD inter-change format}
}
\value{
Dataframe of predictions
}
\description{
Predicts hera model reference indices.
}
\details{
\code{prediction()} predicts model indices.
}
\examples{
predictions <- prediction(demo_data)
}
